requirements.txt
bcrypt==3.1.6
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
Click==7.0
colorize==1.0.4
Django==2.1.7
django-contrib-comments==1.9.1
geojson==2.4.1
gTTS==2.0.3
gTTS-token==1.1.3
idna==2.8
numpy==1.16.1
PTable==0.9.2
PyAudio==0.2.11
pycparser==2.19
pyowm==2.10.0
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.8
speech==0.5.2
SpeechRecognition==3.8.1
urllib3==1.24.1
weather-api==1.0.7
